stick FilterPill to the screen
